[feature]Inputitemの処理をカスタムフック化
